requirements.txt
future==0.16.0
python-whois==0.6.4
requests==2.12.4
